# Audit MemoryExtensions.IndexOf variants