Hi @adamsitnik, after your #73768 I am submitting a few changes which I noticed and presumed were not intentionally left out:

1. Add the `SpanHelpers.Char` specialisation to `LastIndexOf(Span, ROS)` to match `LastIndexOf(ROS, ROS)`. I did not forward the `Span` version to `ROS` version to be consistent with `IndexOf(Span, ROS)` and `IndexOf(ROS, ROS)`, being wary of related discussion in https://github.com/dotnet/runtime/pull/73768#issuecomment-1213152879
2. I did forward the `LastIndexOfAny(Span, ROS)` to `LastIndexOfAny(ROS, ROS)` to be consistent with `IndexOfAny` and considering the next changes:
3. Add more `byte` cases to `LastIndexOfAny` and `IndexOfAny` to be consistent with the `short` path

Things I noticed but did not change, deciding they were intentional:
1. Missing `int-` and `long-`size paths on a few variants. I saw a couple removed in that PR
2. Varying use of `IsBitwiseEquatable` vs `CanVectorizeAndBenefit`. I saw some tweaks in the last commit of that PR
3. `SpanHelpers.IndexOfAnyValueType` has up to 5-value versions but `LastIndexOfAnyValueType` has only up to 4-value

~~Sadly I am having a few troubles getting some benchmarks running. I will submit this as a draft until then, if you see anything I am doing obviously wrong I would appreciate the pointer (I am a rookie with the runtime repo).~~

edit: I was using the wrong corerun path. I changed
`dotnet run -c Release -f net8.0 -- --coreRun C:\src\artifacts\bin\coreclr\windows.x64.Release\corerun.exe C:\src\runtime\artifacts\bin\coreclr\windows.x64.Release\corerun.exe`

to

`dotnet run -c Release -- --coreRun C:\src\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe C:\src\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe`