@rob-hague 2022-09-17
Benchmarks summary:

Regressions in IndexOfAny_byte_ROS and LastIndexOfAny_byte_ROS for 2-value
Regressions in LastIndexOfAny_byte_Span and LastIndexOfAny_char_Span for 5-value

I guess the former is related to codegen.
I guess LastIndexOfAny_byte_Span 5-value suffers from extra indirection to call the same method (there is no vectorized 5-value LastIndexOfAny).
My guess for LastIndexOfAny_char_Span 5-value is that this comes from benchmarking an unrepresentative search-space (and the probabilistic map suffers heavily from this)

I will try a more representative search space soon.

<details>

``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22000.978/21H2)
AMD Ryzen 7 5700U with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.100-alpha.1.22465.14
  [Host]     : .NET 8.0.0 (8.0.22.46215), X64 RyuJIT AVX2
  main : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  PR : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2


```
|                   Method |        Job | SearchSpaceCount | ValueSpaceCount |         Mean |      Error |      StdDev |       Median | Ratio | RatioSD |
|------------------------- |----------- |----------------- |---------------- |-------------:|-----------:|------------:|-------------:|------:|--------:|
|      IndexOfAny_byte_ROS | main |                8 |               1 |     9.830 ns |  0.2249 ns |   0.6156 ns |    10.041 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |                8 |               1 |     4.971 ns |  0.1001 ns |   0.0937 ns |     4.979 ns |  0.55 |    0.04 |
|      IndexOfAny_byte_ROS | main |                8 |               2 |     6.129 ns |  0.1493 ns |   0.2576 ns |     5.950 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |                8 |               2 |     8.934 ns |  0.2029 ns |   0.3957 ns |     8.968 ns |  1.46 |    0.09 |
|      IndexOfAny_byte_ROS | main |                8 |               3 |     9.384 ns |  0.2156 ns |   0.4154 ns |     9.364 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |                8 |               3 |     9.654 ns |  0.2196 ns |   0.5387 ns |     9.661 ns |  1.03 |    0.08 |
|      IndexOfAny_byte_ROS | main |                8 |               4 |    18.289 ns |  0.3919 ns |   0.7067 ns |    18.407 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |                8 |               4 |    11.070 ns |  0.2500 ns |   0.5107 ns |    11.093 ns |  0.61 |    0.04 |
|      IndexOfAny_byte_ROS | main |                8 |               5 |    21.680 ns |  0.4553 ns |   0.5592 ns |    21.706 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |                8 |               5 |    10.146 ns |  0.2242 ns |   0.2399 ns |    10.126 ns |  0.47 |    0.02 |
|      IndexOfAny_byte_ROS | main |               64 |               1 |    40.300 ns |  0.2855 ns |   0.2670 ns |    40.281 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |               64 |               1 |     4.416 ns |  0.1032 ns |   0.0966 ns |     4.435 ns |  0.11 |    0.00 |
|      IndexOfAny_byte_ROS | main |               64 |               2 |     4.104 ns |  0.1106 ns |   0.1786 ns |     4.115 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |               64 |               2 |     5.404 ns |  0.1132 ns |   0.1059 ns |     5.400 ns |  1.32 |    0.07 |
|      IndexOfAny_byte_ROS | main |               64 |               3 |     5.657 ns |  0.1084 ns |   0.1014 ns |     5.677 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |               64 |               3 |     5.975 ns |  0.1438 ns |   0.2196 ns |     6.006 ns |  1.06 |    0.04 |
|      IndexOfAny_byte_ROS | main |               64 |               4 |    98.455 ns |  1.2650 ns |   1.1833 ns |    98.209 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |               64 |               4 |     7.705 ns |  0.1821 ns |   0.4364 ns |     7.683 ns |  0.08 |    0.00 |
|      IndexOfAny_byte_ROS | main |               64 |               5 |   111.305 ns |  0.5146 ns |   0.4814 ns |   111.450 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |               64 |               5 |     8.245 ns |  0.1912 ns |   0.3448 ns |     8.240 ns |  0.07 |    0.00 |
|      IndexOfAny_byte_ROS | main |             1024 |               1 |   550.750 ns |  0.4103 ns |   0.3427 ns |   550.611 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |             1024 |               1 |    16.185 ns |  0.1722 ns |   0.1611 ns |    16.210 ns |  0.03 |    0.00 |
|      IndexOfAny_byte_ROS | main |             1024 |               2 |    18.040 ns |  0.3859 ns |   0.6657 ns |    17.858 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |             1024 |               2 |    20.046 ns |  0.4286 ns |   0.4936 ns |    20.117 ns |  1.11 |    0.05 |
|      IndexOfAny_byte_ROS | main |             1024 |               3 |    22.593 ns |  0.3136 ns |   0.2934 ns |    22.444 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |             1024 |               3 |    22.648 ns |  0.4287 ns |   0.4011 ns |    22.742 ns |  1.00 |    0.03 |
|      IndexOfAny_byte_ROS | main |             1024 |               4 | 1,378.781 ns |  5.7639 ns |   5.3915 ns | 1,379.800 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |             1024 |               4 |    30.109 ns |  0.6085 ns |   0.5692 ns |    30.250 ns |  0.02 |    0.00 |
|      IndexOfAny_byte_ROS | main |             1024 |               5 | 1,648.225 ns |  3.7257 ns |   3.1112 ns | 1,648.891 ns |  1.00 |    0.00 |
|      IndexOfAny_byte_ROS | PR |             1024 |               5 |    30.211 ns |  0.4717 ns |   0.4412 ns |    30.297 ns |  0.02 |    0.00 |
|    LastIndexOf_char_Span | main |                8 |               1 |     4.632 ns |  0.0077 ns |   0.0072 ns |     4.630 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |                8 |               1 |     4.694 ns |  0.1110 ns |   0.1695 ns |     4.716 ns |  1.01 |    0.05 |
|    LastIndexOf_char_Span | main |                8 |               2 |     6.081 ns |  0.0035 ns |   0.0029 ns |     6.080 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |                8 |               2 |     6.635 ns |  0.0105 ns |   0.0088 ns |     6.637 ns |  1.09 |    0.00 |
|    LastIndexOf_char_Span | main |                8 |               3 |     5.761 ns |  0.0049 ns |   0.0046 ns |     5.763 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |                8 |               3 |     6.268 ns |  0.0124 ns |   0.0104 ns |     6.266 ns |  1.09 |    0.00 |
|    LastIndexOf_char_Span | main |                8 |               4 |     5.221 ns |  0.0063 ns |   0.0059 ns |     5.220 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |                8 |               4 |     6.004 ns |  0.0035 ns |   0.0032 ns |     6.005 ns |  1.15 |    0.00 |
|    LastIndexOf_char_Span | main |                8 |               5 |     4.617 ns |  0.0015 ns |   0.0011 ns |     4.617 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |                8 |               5 |     5.771 ns |  0.0028 ns |   0.0025 ns |     5.770 ns |  1.25 |    0.00 |
|    LastIndexOf_char_Span | main |               64 |               1 |    17.617 ns |  0.0112 ns |   0.0099 ns |    17.616 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |               64 |               1 |     5.610 ns |  0.0049 ns |   0.0043 ns |     5.611 ns |  0.32 |    0.00 |
|    LastIndexOf_char_Span | main |               64 |               2 |    19.020 ns |  0.0119 ns |   0.0112 ns |    19.022 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |               64 |               2 |     7.238 ns |  0.0725 ns |   0.0642 ns |     7.253 ns |  0.38 |    0.00 |
|    LastIndexOf_char_Span | main |               64 |               3 |    18.526 ns |  0.0131 ns |   0.0110 ns |    18.524 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |               64 |               3 |     7.215 ns |  0.0510 ns |   0.0477 ns |     7.188 ns |  0.39 |    0.00 |
|    LastIndexOf_char_Span | main |               64 |               4 |    18.194 ns |  0.0127 ns |   0.0113 ns |    18.192 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |               64 |               4 |     7.288 ns |  0.0711 ns |   0.0665 ns |     7.296 ns |  0.40 |    0.00 |
|    LastIndexOf_char_Span | main |               64 |               5 |    18.095 ns |  0.0253 ns |   0.0237 ns |    18.097 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |               64 |               5 |     7.137 ns |  0.0529 ns |   0.0495 ns |     7.142 ns |  0.39 |    0.00 |
|    LastIndexOf_char_Span | main |             1024 |               1 |   220.402 ns |  0.2264 ns |   0.2118 ns |   220.303 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |             1024 |               1 |    29.481 ns |  0.0918 ns |   0.0858 ns |    29.499 ns |  0.13 |    0.00 |
|    LastIndexOf_char_Span | main |             1024 |               2 |   219.591 ns |  0.3893 ns |   0.3451 ns |   219.559 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |             1024 |               2 |    59.159 ns |  0.7284 ns |   0.6813 ns |    59.171 ns |  0.27 |    0.00 |
|    LastIndexOf_char_Span | main |             1024 |               3 |   218.000 ns |  0.1500 ns |   0.1403 ns |   217.961 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |             1024 |               3 |    59.667 ns |  0.3749 ns |   0.3323 ns |    59.617 ns |  0.27 |    0.00 |
|    LastIndexOf_char_Span | main |             1024 |               4 |   220.605 ns |  0.0831 ns |   0.0737 ns |   220.620 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |             1024 |               4 |    59.760 ns |  0.4245 ns |   0.3763 ns |    59.789 ns |  0.27 |    0.00 |
|    LastIndexOf_char_Span | main |             1024 |               5 |   219.007 ns |  0.1947 ns |   0.1626 ns |   219.009 ns |  1.00 |    0.00 |
|    LastIndexOf_char_Span | PR |             1024 |               5 |    59.505 ns |  0.5208 ns |   0.4871 ns |    59.411 ns |  0.27 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |                8 |               1 |     8.740 ns |  0.1974 ns |   0.2027 ns |     8.756 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |                8 |               1 |     5.412 ns |  0.1336 ns |   0.1641 ns |     5.431 ns |  0.62 |    0.02 |
|  LastIndexOfAny_byte_ROS | main |                8 |               2 |     6.840 ns |  0.1583 ns |   0.2059 ns |     6.875 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |                8 |               2 |     8.224 ns |  0.1198 ns |   0.1062 ns |     8.235 ns |  1.21 |    0.05 |
|  LastIndexOfAny_byte_ROS | main |                8 |               3 |     9.525 ns |  0.2188 ns |   0.3341 ns |     9.553 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |                8 |               3 |     9.984 ns |  0.2122 ns |   0.1985 ns |     9.950 ns |  1.05 |    0.05 |
|  LastIndexOfAny_byte_ROS | main |                8 |               4 |    16.621 ns |  0.3242 ns |   0.3033 ns |    16.816 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |                8 |               4 |    11.422 ns |  0.2458 ns |   0.3108 ns |    11.439 ns |  0.69 |    0.02 |
|  LastIndexOfAny_byte_ROS | main |                8 |               5 |    17.627 ns |  0.1521 ns |   0.1422 ns |    17.653 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |                8 |               5 |    18.195 ns |  0.1369 ns |   0.1281 ns |    18.194 ns |  1.03 |    0.01 |
|  LastIndexOfAny_byte_ROS | main |               64 |               1 |    38.750 ns |  0.2145 ns |   0.2006 ns |    38.723 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |               64 |               1 |     4.607 ns |  0.1005 ns |   0.0891 ns |     4.613 ns |  0.12 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |               64 |               2 |     4.612 ns |  0.1186 ns |   0.1584 ns |     4.656 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |               64 |               2 |     5.156 ns |  0.0731 ns |   0.0684 ns |     5.193 ns |  1.12 |    0.04 |
|  LastIndexOfAny_byte_ROS | main |               64 |               3 |     5.537 ns |  0.1178 ns |   0.1102 ns |     5.542 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |               64 |               3 |     6.040 ns |  0.1475 ns |   0.3012 ns |     6.163 ns |  1.08 |    0.07 |
|  LastIndexOfAny_byte_ROS | main |               64 |               4 |    98.505 ns |  0.6669 ns |   0.6238 ns |    98.666 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |               64 |               4 |     7.229 ns |  0.1715 ns |   0.3424 ns |     7.249 ns |  0.07 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |               64 |               5 |   109.523 ns |  0.1604 ns |   0.1500 ns |   109.560 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |               64 |               5 |   111.629 ns |  0.1737 ns |   0.1624 ns |   111.664 ns |  1.02 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |             1024 |               1 |   548.149 ns |  0.4610 ns |   0.4313 ns |   548.020 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |             1024 |               1 |    17.031 ns |  0.0456 ns |   0.0426 ns |    17.028 ns |  0.03 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |             1024 |               2 |    20.863 ns |  0.3697 ns |   0.3458 ns |    20.693 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |             1024 |               2 |    18.833 ns |  0.2013 ns |   0.1883 ns |    18.859 ns |  0.90 |    0.02 |
|  LastIndexOfAny_byte_ROS | main |             1024 |               3 |    21.567 ns |  0.2792 ns |   0.2612 ns |    21.573 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |             1024 |               3 |    21.688 ns |  0.4454 ns |   0.4766 ns |    21.688 ns |  1.00 |    0.03 |
|  LastIndexOfAny_byte_ROS | main |             1024 |               4 | 1,371.158 ns |  8.6512 ns |   8.0923 ns | 1,372.419 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |             1024 |               4 |    29.946 ns |  0.3665 ns |   0.3428 ns |    30.049 ns |  0.02 |    0.00 |
|  LastIndexOfAny_byte_ROS | main |             1024 |               5 | 1,644.423 ns |  2.4255 ns |   2.2688 ns | 1,644.547 ns |  1.00 |    0.00 |
|  LastIndexOfAny_byte_ROS | PR |             1024 |               5 | 1,691.436 ns | 39.6828 ns | 105.9215 ns | 1,646.512 ns |  1.07 |    0.11 |
| LastIndexOfAny_byte_Span | main |                8 |               1 |     7.177 ns |  0.0663 ns |   0.0621 ns |     7.179 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |                8 |               1 |     5.121 ns |  0.0870 ns |   0.0771 ns |     5.122 ns |  0.71 |    0.02 |
| LastIndexOfAny_byte_Span | main |                8 |               2 |    10.563 ns |  0.1535 ns |   0.1436 ns |    10.639 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |                8 |               2 |     9.301 ns |  0.2136 ns |   0.4012 ns |     9.432 ns |  0.88 |    0.04 |
| LastIndexOfAny_byte_Span | main |                8 |               3 |    13.617 ns |  0.2393 ns |   0.2238 ns |    13.631 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |                8 |               3 |    10.058 ns |  0.1335 ns |   0.1183 ns |    10.039 ns |  0.74 |    0.01 |
| LastIndexOfAny_byte_Span | main |                8 |               4 |    16.143 ns |  0.2179 ns |   0.2038 ns |    16.174 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |                8 |               4 |    12.045 ns |  0.1427 ns |   0.1265 ns |    12.010 ns |  0.75 |    0.01 |
| LastIndexOfAny_byte_Span | main |                8 |               5 |    17.577 ns |  0.1953 ns |   0.1631 ns |    17.554 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |                8 |               5 |    26.138 ns |  0.7270 ns |   2.1437 ns |    27.469 ns |  1.45 |    0.12 |
| LastIndexOfAny_byte_Span | main |               64 |               1 |    38.526 ns |  0.1101 ns |   0.1030 ns |    38.559 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |               64 |               1 |     4.531 ns |  0.0941 ns |   0.0880 ns |     4.552 ns |  0.12 |    0.00 |
| LastIndexOfAny_byte_Span | main |               64 |               2 |    59.828 ns |  0.1546 ns |   0.1291 ns |    59.875 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |               64 |               2 |     5.825 ns |  0.1112 ns |   0.1041 ns |     5.802 ns |  0.10 |    0.00 |
| LastIndexOfAny_byte_Span | main |               64 |               3 |    78.977 ns |  1.6081 ns |   1.7207 ns |    78.920 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |               64 |               3 |     6.372 ns |  0.1522 ns |   0.2665 ns |     6.466 ns |  0.08 |    0.00 |
| LastIndexOfAny_byte_Span | main |               64 |               4 |    92.914 ns |  0.2358 ns |   0.2206 ns |    92.887 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |               64 |               4 |     7.106 ns |  0.1092 ns |   0.1021 ns |     7.093 ns |  0.08 |    0.00 |
| LastIndexOfAny_byte_Span | main |               64 |               5 |   115.788 ns |  0.4738 ns |   0.4200 ns |   115.776 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |               64 |               5 |   178.753 ns |  2.8791 ns |   2.4042 ns |   178.450 ns |  1.54 |    0.02 |
| LastIndexOfAny_byte_Span | main |             1024 |               1 |   546.179 ns |  0.7525 ns |   0.7039 ns |   546.434 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |             1024 |               1 |    15.804 ns |  0.2782 ns |   0.2603 ns |    15.831 ns |  0.03 |    0.00 |
| LastIndexOfAny_byte_Span | main |             1024 |               2 |   826.460 ns |  0.4934 ns |   0.4374 ns |   826.339 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |             1024 |               2 |    22.172 ns |  0.4427 ns |   0.4546 ns |    22.202 ns |  0.03 |    0.00 |
| LastIndexOfAny_byte_Span | main |             1024 |               3 | 1,116.313 ns | 13.1206 ns |  12.2730 ns | 1,118.085 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |             1024 |               3 |    23.311 ns |  0.4563 ns |   0.4268 ns |    23.179 ns |  0.02 |    0.00 |
| LastIndexOfAny_byte_Span | main |             1024 |               4 | 1,377.804 ns |  8.4186 ns |   7.8747 ns | 1,377.064 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |             1024 |               4 |    30.467 ns |  0.1921 ns |   0.1703 ns |    30.499 ns |  0.02 |    0.00 |
| LastIndexOfAny_byte_Span | main |             1024 |               5 | 1,639.299 ns |  3.6586 ns |   3.4223 ns | 1,639.443 ns |  1.00 |    0.00 |
| LastIndexOfAny_byte_Span | PR |             1024 |               5 | 1,639.896 ns |  2.3544 ns |   2.0871 ns | 1,639.982 ns |  1.00 |    0.00 |
|  LastIndexOfAny_char_ROS | main |                8 |               1 |     4.176 ns |  0.1122 ns |   0.2753 ns |     4.304 ns |  1.00 |    0.00 |
|  LastIndexOfAny_char_ROS | PR |                8 |               1 |     4.118 ns |  0.1111 ns |   0.2908 ns |     4.198 ns |  0.99 |    0.09 |
|  LastIndexOfAny_char_ROS | main |                8 |               2 |     4.153 ns |  0.1100 ns |   0.1391 ns |     4.225 ns |  1.00 |    0.00 |
|  LastIndexOfAny_char_ROS | PR |                8 |               2 |     4.123 ns |  0.1131 ns |   0.2336 ns |     4.225 ns |  1.00 |    0.06 |
|  LastIndexOfAny_char_ROS | main |                8 |               3 |     5.568 ns |  0.1368 ns |   0.1827 ns |     5.593 ns |  1.00 |    0.00 |
|  LastIndexOfAny_char_ROS | PR |                8 |               3 |     5.579 ns |  0.1354 ns |   0.1854 ns |     5.606 ns |  1.00 |    0.05 |
…